backend/requirements.txt
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
cffi==2.0.0
contextlib2==21.6.0
cryptography==46.0.6
Django==6.0.3
django-ninja==1.6.2
django-ninja-extra==0.31.4
django-ninja-jwt==5.4.4
dnspython==2.8.0
email-validator==2.3.0
idna==3.11
injector==0.24.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
